move buttons in atoms directory
