+++ src/components/atomic/buttons/DavitButton.css
@@ -1,0 +1,13 @@
+button {
+    color: var(--text-color);
+    border-radius: var(--davit-border-radius);
+    border: var(--davit-border-size-medium) solid var(--color-secondary);
+    background-color: transparent;
+    text-align: center;
+    padding: var(--davit-padding-top-bottom) var(--davit-padding-right-left);
+}
+
+button:hover {
+    background: var(--davit-hover-color);
+    cursor: pointer;
+}
