+++ src/components/atomic/buttons/DavitDownloadButton.tsx
@@ -1,6 +1,6 @@
 import { faDownload } from "@fortawesome/free-solid-svg-icons";
 import React, { FunctionComponent, useState } from "react";
-import { DavitDownloadModal } from "../modals/DavitDownlaodModal";
+import { DavitDownloadModal } from "../../common/fragments/modals/DavitDownlaodModal";
 import { DavitButton } from "./DavitButton";
 
 export interface DavitDownloadButtonProps {
