+++ src/components/controllPanel/presentation/fragments/edit/fragments/forms/ChainDecisionForm.tsx
@@ -2,14 +2,14 @@ import React, { FunctionComponent } from "react";
 import { ConditionTO } from "../../../../../../../dataAccess/access/to/ConditionTO";
 import { GoToTypesChain } from "../../../../../../../dataAccess/access/types/GoToTypeChain";
 import { DavitUtil } from "../../../../../../../utils/DavitUtil";
+import { DavitAddButton } from "../../../../../../atomic/buttons/DavitAddButton";
+import { DavitBackButton } from "../../../../../../atomic/buttons/DavitBackButton";
+import { DavitDeleteButton } from "../../../../../../atomic/buttons/DavitDeleteButton";
 import { ActorDropDown } from "../../../../../../atomic/dropdowns/ActorDropDown";
 import { ChainDecisionDropDown } from "../../../../../../atomic/dropdowns/ChainDecisionDropDown";
 import { ChainLinkDropDown } from "../../../../../../atomic/dropdowns/ChainLinkDropDown";
 import { GoToChainOptionDropDown } from "../../../../../../atomic/dropdowns/GoToChainOptionDropDown";
 import { InstanceDropDown } from "../../../../../../atomic/dropdowns/InstanceDropDown";
-import { DavitAddButton } from "../../../../../../common/fragments/buttons/DavitAddButton";
-import { DavitBackButton } from "../../../../../../common/fragments/buttons/DavitBackButton";
-import { DavitDeleteButton } from "../../../../../../common/fragments/buttons/DavitDeleteButton";
 import { DavitLabelTextfield } from "../../../../../../common/fragments/DavitLabelTextfield";
 import { Form } from "../../../../../../common/fragments/forms/Form";
 import { FormBody } from "../../../../../../common/fragments/forms/FormBody";
