+++ src/components/controllPanel/presentation/fragments/edit/fragments/forms/DecisionForm.tsx
@@ -2,16 +2,16 @@ import React, { FunctionComponent } from "react";
 import { ConditionTO } from "../../../../../../../dataAccess/access/to/ConditionTO";
 import { GoToTypes } from "../../../../../../../dataAccess/access/types/GoToType";
 import { DavitUtil } from "../../../../../../../utils/DavitUtil";
+import { DavitAddButton } from "../../../../../../atomic/buttons/DavitAddButton";
+import { DavitBackButton } from "../../../../../../atomic/buttons/DavitBackButton";
+import { DavitCommentButton } from "../../../../../../atomic/buttons/DavitCommentButton";
+import { DavitDeleteButton } from "../../../../../../atomic/buttons/DavitDeleteButton";
+import { DavitRootButton } from "../../../../../../atomic/buttons/DavitRootButton";
 import { ActorDropDown } from "../../../../../../atomic/dropdowns/ActorDropDown";
 import { DecisionDropDown } from "../../../../../../atomic/dropdowns/DecisionDropDown";
 import { GoToOptionDropDown } from "../../../../../../atomic/dropdowns/GoToOptionDropDown";
 import { InstanceDropDown } from "../../../../../../atomic/dropdowns/InstanceDropDown";
 import { StepDropDown } from "../../../../../../atomic/dropdowns/StepDropDown";
-import { DavitAddButton } from "../../../../../../common/fragments/buttons/DavitAddButton";
-import { DavitBackButton } from "../../../../../../common/fragments/buttons/DavitBackButton";
-import { DavitCommentButton } from "../../../../../../common/fragments/buttons/DavitCommentButton";
-import { DavitDeleteButton } from "../../../../../../common/fragments/buttons/DavitDeleteButton";
-import { DavitRootButton } from "../../../../../../common/fragments/buttons/DavitRootButton";
 import { DavitLabelTextfield } from "../../../../../../common/fragments/DavitLabelTextfield";
 import { Form } from "../../../../../../common/fragments/forms/Form";
 import { FormBody } from "../../../../../../common/fragments/forms/FormBody";
