+++ src/components/controllPanel/presentation/fragments/edit/fragments/forms/StepForm.tsx
@@ -3,17 +3,17 @@ import { faAngleUp } from "@fortawesome/free-solid-svg-icons/faAngleUp";
 import React, { FunctionComponent, useState } from "react";
 import { ActionTO } from "../../../../../../../dataAccess/access/to/ActionTO";
 import { GoToTypes } from "../../../../../../../dataAccess/access/types/GoToType";
+import { DavitAddButton } from "../../../../../../atomic/buttons/DavitAddButton";
+import { DavitBackButton } from "../../../../../../atomic/buttons/DavitBackButton";
+import { DavitButton } from "../../../../../../atomic/buttons/DavitButton";
+import { DavitCommentButton } from "../../../../../../atomic/buttons/DavitCommentButton";
+import { DavitDeleteButton } from "../../../../../../atomic/buttons/DavitDeleteButton";
+import { DavitEditButton } from "../../../../../../atomic/buttons/DavitEditButton";
+import { DavitRootButton } from "../../../../../../atomic/buttons/DavitRootButton";
+import { DavitShowMoreButton } from "../../../../../../atomic/buttons/DavitShowMoreButton";
 import { DecisionDropDown } from "../../../../../../atomic/dropdowns/DecisionDropDown";
 import { GoToOptionDropDown } from "../../../../../../atomic/dropdowns/GoToOptionDropDown";
 import { StepDropDown } from "../../../../../../atomic/dropdowns/StepDropDown";
-import { DavitAddButton } from "../../../../../../common/fragments/buttons/DavitAddButton";
-import { DavitBackButton } from "../../../../../../common/fragments/buttons/DavitBackButton";
-import { DavitButton } from "../../../../../../common/fragments/buttons/DavitButton";
-import { DavitCommentButton } from "../../../../../../common/fragments/buttons/DavitCommentButton";
-import { DavitDeleteButton } from "../../../../../../common/fragments/buttons/DavitDeleteButton";
-import { DavitEditButton } from "../../../../../../common/fragments/buttons/DavitEditButton";
-import { DavitRootButton } from "../../../../../../common/fragments/buttons/DavitRootButton";
-import { DavitShowMoreButton } from "../../../../../../common/fragments/buttons/DavitShowMoreButton";
 import { DavitLabelTextfield } from "../../../../../../common/fragments/DavitLabelTextfield";
 import { Form } from "../../../../../../common/fragments/forms/Form";
 import { FormBody } from "../../../../../../common/fragments/forms/FormBody";
