+++ src/components/controllPanel/presentation/fragments/file/ControlPanelFileController.tsx
@@ -2,9 +2,9 @@ import React, { FunctionComponent } from "react";
 import { useDispatch } from "react-redux";
 import { EditActions } from "../../../../../slices/EditSlice";
 import { GlobalActions } from "../../../../../slices/GlobalSlice";
-import { DavitDeleteButton } from "../../../../common/fragments/buttons/DavitDeleteButton";
-import { DavitDownloadButton } from "../../../../common/fragments/buttons/DavitDownloadButton";
-import { DavitUploadButton } from "../../../../common/fragments/buttons/DavitUploadButton";
+import { DavitDeleteButton } from "../../../../atomic/buttons/DavitDeleteButton";
+import { DavitDownloadButton } from "../../../../atomic/buttons/DavitDownloadButton";
+import { DavitUploadButton } from "../../../../atomic/buttons/DavitUploadButton";
 import { ControlPanel } from "../edit/common/ControlPanel";
 import { OptionField } from "../edit/common/OptionField";
