+++ src/components/controllPanel/presentation/fragments/tabs/ControlPanelTabController.tsx
@@ -1,7 +1,7 @@
 import { faShare } from "@fortawesome/free-solid-svg-icons";
 import React, { FunctionComponent } from "react";
 import { ModuleRoutes } from "../../../../../pages/Davit";
-import { DavitButton } from "../../../../common/fragments/buttons/DavitButton";
+import { DavitButton } from "../../../../atomic/buttons/DavitButton";
 import { ControlPanel } from "../edit/common/ControlPanel";
 import { OptionField } from "../edit/common/OptionField";
